+++ aoc/src/2018/day-9.cpp
@@ -15,8 +15,8 @@ struct advent_2018_9 : problem
 	}
 
 protected:
-	std::size_t players = 0;
-	std::size_t score = 0;
+	std::uintmax_t players = 0;
+	std::uintmax_t score = 0;
 
 	void prepare_input(std::ifstream& fin) override {
 		std::string line;
@@ -28,22 +28,22 @@ protected:
 		score = std::stoull(m[2]);
 	}
 
-	std::size_t play_game(std::size_t max_players, std::size_t last_score) {
-		std::list<std::size_t> circle;
+	std::size_t play_game(std::uintmax_t max_players, std::uintmax_t last_score) {
+		std::list<std::uintmax_t> circle;
 		circle.push_back(0);
 
 		std::vector<std::size_t> scores(max_players);
 
-		using cl_it = utility::cyclic_iterator<std::list<std::size_t>::iterator>;
+		using cl_it = utility::cyclic_iterator<std::list<std::uintmax_t>::iterator>;
 		cl_it current(std::begin(circle), std::end(circle));
 
-		for(std::size_t marble = 1; marble < last_score; ++marble) {
+		for(std::uintmax_t marble = 1; marble < last_score; ++marble) {
 			if(marble % 23 != 0) {
 				std::advance(current, 2);
 				current = cl_it(circle.insert(current.base(), marble), std::begin(circle), std::end(circle));
 			} else {
 				std::advance(current, -7);
-				const std::size_t removed = *current;
+				const std::uintmax_t removed = *current;
 				current = cl_it(circle.erase(current.base()), std::begin(circle), std::end(circle));
 				scores[marble % max_players] += marble + removed;
 			}
